// src/app/core/services/models/response/TaskResponse.ts
export interface TaskResponse {
  descriptionCreation?: string;
  descriptionUpdate?: string;
  title?: string;
  status?: string;
  dateFin?: string;
  idAffected?: number;
  dateDebut?: string;
  rapport?: string;
  id: number;
  publicId?: string;
  createdDate?: string;
  createdBy?: string;
  lastModifiedDate?: string;
  lastModifiedBy?: string;
}